claudia_rdg_latest/RDG_Chat/requirements.txt
# Azure AI SDKs
azure-identity==1.15.0
azure-core==1.36.0
azure-search-documents==11.6.0
openai==1.109.1
semantic-kernel==1.39.3
azure-cosmos>=4.7.0

# Embedding models for multi-vector search
sentence-transformers>=2.2.0
torch>=1.13.0

# HTTP & API
httpx==0.26.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
aiohttp>=3.9.0
requests>=2.31.0

# Microsoft Bot Framework
botbuilder-core>=4.14.0
botbuilder-schema>=4.14.0
botbuilder-integration-aiohttp>=4.14.0

# Data validation
pydantic>=2.0.0
pydantic[email]>=2.0.0

# Utilities
python-dotenv==1.0.0

# Monitoring (optional)
opencensus-ext-azure==1.1.13
